Rename single-word components to multi-word
Web components should have multiple words, since future HTML (and SVG,
MathML, etc.) elements are guaranteed to be only one word (no hyphens).

As a workaround, experiments with only one word in their file name
should end in "Experiment", but that will be stripped out when displayed
to the user.