eslint: Enforce SFC tag order: <script>, <template>, <style>
This also changes the `<script>` tag like this:

```diff
-<script lang="ts" setup></script>
+<script setup lang="ts"></script>
```

It's debatable whether this is better than `<template>`, `<script>`,
`<style>`, but now we have a rule for it.

What's nice about this order is that it brings the style closer to the
template, while also having the template close to the script. The
previous `<template>`, `<script>`, `<style>` order would de-emphasize
the logic part in place of the visual part.

It also seems that the newer documentation regarding `<script setup>`
uses this order.