Upgrade @highlightjs/vue-plugin to 2.1.2
This fixes an issue with the dependency resolver not finding the correct
file. The commonJS one had an external reference to Vue, which doesn't
exist in this project, and was preventing tests with references to
`highlight.js` from even running. The ESM version doesn't, so it works
if it were specially aliased. Fortunately, this was fixed upstream here:
https://github.com/highlightjs/vue-plugin/pull/27