Update @vitest/coverage-c8 and postcss-dir-pseudo-class
Holding back on vitest for now, since there seems to be a problem with
snapshots not getting tested correctly. When RecursionExperiment.spec.ts
is run explicitly, the snapshot is tested, but when run together with
the rest of the tests, it's marked as "obsolete". Changing
`test.isolated` doesn't affect it.