Upgrade postcss-logical to 6.0.1
This has a breaking change in that it no longer supports the "preserve"
option. This means that it will always replace logical properties with
their positional equivalents, so we just let it do that for now.

Perhaps there may be a way to enable that only on legacy builds, but for
now it doesn't seem likely.