ci: Replace actions-rs/toolchain with dtolnay/rust-toolchain
David Tolnay is the author of some pretty significant creates, such as
`syn`, `thiserror` and `anyhow`, so I think this action will be pretty
reliable, even if it's not an official replacement.

The reason for updating in the first place is that the previous one had
deprecation messages in the logs, and hasn't been updated in ~3 years.