Replace import.meta.glob with constant file list in RecursionExperiment
This actually changes a lot about the build. It turns out that
`import.meta.glob` affects the bundle process quite a bit. By replacing
it with a constant list, a lot of components got inlined with their
views.

I never aimed to reduce the total bundle size, but this actually reduces
the bundle by a couple hundred kB.