Add a way for static variables to be replaced with raw values
This allows for those variables to be minified. For example, if I have
this design token structure and CSS:

```js
export default {
  base: {
    color: {
      $type: "color",
      purple: { $value: "rebeccapurple" },
    },
  },
};
```

```css
.foo {
  color: var(--pv-base-color-purple);
}
```

The CSS will be converted into this:

```css
.foo {
  color: rebeccapurple;
}
```

Which gets minified into this:

```css
.foo{color:#639}
```

Currently only a select number of values are replaced, since they're the
most common and simplest to convert. I'm not sure yet how to convert
fonts or other values, so I'll leave those alone for now.